Refactor cmd/dsl files
Split prototyping.go file into multiple files that are
(hopefully) more descriptive.
